+++ cmd/nlb_service_add.go
@@ -4,16 +4,12 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/spf13/cobra"
 
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
-	"github.com/exoscale/cli/utils"
-	egoscale "github.com/exoscale/egoscale/v2"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type nlbServiceAddCmd struct {
@@ -24,20 +20,20 @@ type nlbServiceAddCmd struct {
 	NetworkLoadBalancer string `cli-arg:"#" cli-usage:"LOAD-BALANCER-NAME|ID"`
 	Name                string `cli-arg:"#" cli-usage:"SERVICE-NAME"`
 
-	Description         string `cli-usage:"service description"`
-	HealthcheckInterval int64  `cli-usage:"service health checking interval in seconds"`
-	HealthcheckMode     string `cli-usage:"service health checking mode (tcp|http|https)"`
-	HealthcheckPort     int64  `cli-usage:"service health checking port (defaults to target port)"`
-	HealthcheckRetries  int64  `cli-usage:"service health checking retries"`
-	HealthcheckTLSSNI   string `cli-flag:"healthcheck-tls-sni" cli-usage:"service health checking server name to present with SNI in https mode"`
-	HealthcheckTimeout  int64  `cli-usage:"service health checking timeout in seconds"`
-	HealthcheckURI      string `cli-usage:"service health checking URI (required in http(s) mode)"`
-	InstancePool        string `cli-usage:"name or ID of the Instance Pool to forward traffic to"`
-	Port                int64  `cli-usage:"service port"`
-	Protocol            string `cli-usage:"service network protocol (tcp|udp)"`
-	Strategy            string `cli-usage:"load balancing strategy (round-robin|source-hash)"`
-	TargetPort          int64  `cli-usage:"port to forward traffic to on target instances (defaults to service port)"`
-	Zone                string `cli-short:"z" cli-usage:"Network Load Balancer zone"`
+	Description         string      `cli-usage:"service description"`
+	HealthcheckInterval int64       `cli-usage:"service health checking interval in seconds"`
+	HealthcheckMode     string      `cli-usage:"service health checking mode (tcp|http|https)"`
+	HealthcheckPort     int64       `cli-usage:"service health checking port (defaults to target port)"`
+	HealthcheckRetries  int64       `cli-usage:"service health checking retries"`
+	HealthcheckTLSSNI   string      `cli-flag:"healthcheck-tls-sni" cli-usage:"service health checking server name to present with SNI in https mode"`
+	HealthcheckTimeout  int64       `cli-usage:"service health checking timeout in seconds"`
+	HealthcheckURI      string      `cli-usage:"service health checking URI (required in http(s) mode)"`
+	InstancePool        string      `cli-usage:"name or ID of the Instance Pool to forward traffic to"`
+	Port                int64       `cli-usage:"service port"`
+	Protocol            string      `cli-usage:"service network protocol (tcp|udp)"`
+	Strategy            string      `cli-usage:"load balancing strategy (round-robin|source-hash)"`
+	TargetPort          int64       `cli-usage:"port to forward traffic to on target instances (defaults to service port)"`
+	Zone                v3.ZoneName `cli-short:"z" cli-usage:"Network Load Balancer zone"`
 }
 
 func (c *nlbServiceAddCmd) cmdAliases() []string { return nil }
@@ -57,58 +53,75 @@ func (c *nlbServiceAddCmd) cmdPreRun(cmd *cobra.Command, args []string) error {
 }
 
 func (c *nlbServiceAddCmd) cmdRun(_ *cobra.Command, _ []string) error {
-	var (
-		port       = uint16(c.Port)
-		targetPort = uint16(c.TargetPort)
-		hcPort     = uint16(c.HealthcheckPort)
-		hcInterval = time.Duration(c.HealthcheckInterval) * time.Second
-		hcTimeout  = time.Duration(c.HealthcheckTimeout) * time.Second
-	)
 
-	service := &egoscale.NetworkLoadBalancerService{
-		Description: utils.NonEmptyStringPtr(c.Description),
-		Healthcheck: &egoscale.NetworkLoadBalancerServiceHealthcheck{
-			Interval: &hcInterval,
-			Mode:     &c.HealthcheckMode,
-			Port:     &hcPort,
-			Retries:  &c.HealthcheckRetries,
-			TLSSNI:   utils.NonEmptyStringPtr(c.HealthcheckTLSSNI),
-			Timeout:  &hcTimeout,
-			URI:      utils.NonEmptyStringPtr(c.HealthcheckURI),
+	service := v3.AddServiceToLoadBalancerRequest{
+		Description: c.Description,
+		Healthcheck: &v3.LoadBalancerServiceHealthcheck{
+			Interval: c.HealthcheckInterval,
+			Mode:     v3.LoadBalancerServiceHealthcheckMode(c.HealthcheckMode),
+			Retries:  c.HealthcheckRetries,
+			Timeout:  c.HealthcheckTimeout,
 		},
-		Name:       &c.Name,
-		Port:       &port,
-		Protocol:   &c.Protocol,
-		Strategy:   &c.Strategy,
-		TargetPort: &targetPort,
+		Name:         c.Name,
+		Port:         c.Port,
+		Protocol:     v3.AddServiceToLoadBalancerRequestProtocol(c.Protocol),
+		Strategy:     v3.AddServiceToLoadBalancerRequestStrategy(c.Strategy),
+		TargetPort:   c.TargetPort,
+		InstancePool: &v3.InstancePool{},
 	}
 
-	if strings.HasPrefix(*service.Healthcheck.Mode, "http") && *service.Healthcheck.URI == "" {
-		return errors.New(`an healthcheck URI is required in "http(s)" mode`)
+	ctx := gContext
+
+	client, err := switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, c.Zone)
+	if err != nil {
+		return err
 	}
 
-	if *service.TargetPort == 0 {
+	if strings.HasPrefix(string(service.Healthcheck.Mode), "http") && c.HealthcheckURI == "" {
+		return errors.New(`a healthcheck URI is required in "http(s)" mode`)
+	}
+
+	if service.TargetPort == 0 {
 		service.TargetPort = service.Port
 	}
-	if *service.Healthcheck.Port == 0 {
+	if service.Healthcheck.Port == 0 {
 		service.Healthcheck.Port = service.TargetPort
 	}
 
-	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
-
-	nlb, err := globalstate.EgoscaleClient.FindNetworkLoadBalancer(ctx, c.Zone, c.NetworkLoadBalancer)
-	if err != nil {
-		return fmt.Errorf("error retrieving Network Load Balancer: %w", err)
+	if c.HealthcheckURI != "" {
+		service.Healthcheck.URI = c.HealthcheckURI
+	}
+	if c.HealthcheckTLSSNI != "" {
+		service.Healthcheck.TlsSNI = c.HealthcheckTLSSNI
 	}
 
-	instancePool, err := globalstate.EgoscaleClient.FindInstancePool(ctx, c.Zone, c.InstancePool)
+	nlbs, err := client.ListLoadBalancers(ctx)
 	if err != nil {
-		return fmt.Errorf("error retrieving Instance Pool: %w", err)
+		return err
 	}
-	service.InstancePoolID = instancePool.ID
+	nlb, err := nlbs.FindLoadBalancer(c.NetworkLoadBalancer)
+	if err != nil {
+		return err
+	}
+
+	instancePools, err := client.ListInstancePools(ctx)
+	if err != nil {
+		return err
+	}
+
+	instancePool, err := instancePools.FindInstancePool(c.InstancePool)
+	if err != nil {
+		return err
+	}
+	service.InstancePool.ID = instancePool.ID
+
+	op, err := client.AddServiceToLoadBalancer(ctx, nlb.ID, service)
+	if err != nil {
+		return err
+	}
 
 	decorateAsyncOperation(fmt.Sprintf("Adding service %q...", c.Name), func() {
-		service, err = globalstate.EgoscaleClient.CreateNetworkLoadBalancerService(ctx, c.Zone, nlb, service)
+		op, err = client.Wait(ctx, op, v3.OperationStateSuccess)
 	})
 	if err != nil {
 		return err
@@ -117,8 +130,8 @@ func (c *nlbServiceAddCmd) cmdRun(_ *cobra.Command, _ []string) error {
 	if !globalstate.Quiet {
 		return (&nlbServiceShowCmd{
 			cliCommandSettings:  c.cliCommandSettings,
-			NetworkLoadBalancer: *nlb.ID,
-			Service:             *service.ID,
+			NetworkLoadBalancer: nlb.ID.String(),
+			Service:             service.Name,
 			Zone:                c.Zone,
 		}).cmdRun(nil, nil)
 	}