+++ cmd/nlb_service_delete.go
@@ -6,9 +6,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type nlbServiceDeleteCmd struct {
@@ -19,8 +18,8 @@ type nlbServiceDeleteCmd struct {
 	NetworkLoadBalancer string `cli-arg:"#" cli-usage:"LOAD-BALANCER-NAME|ID"`
 	Service             string `cli-arg:"#" cli-usage:"SERVICE-NAME|ID"`
 
-	Force bool   `cli-short:"f" cli-usage:"don't prompt for confirmation"`
-	Zone  string `cli-short:"z" cli-usage:"Network Load Balancer zone"`
+	Force bool        `cli-short:"f" cli-usage:"don't prompt for confirmation"`
+	Zone  v3.ZoneName `cli-short:"z" cli-usage:"Network Load Balancer zone"`
 }
 
 func (c *nlbServiceDeleteCmd) cmdAliases() []string { return gRemoveAlias }
@@ -35,30 +34,36 @@ func (c *nlbServiceDeleteCmd) cmdPreRun(cmd *cobra.Command, args []string) error
 }
 
 func (c *nlbServiceDeleteCmd) cmdRun(_ *cobra.Command, _ []string) error {
-	if !c.Force {
-		if !askQuestion(fmt.Sprintf("Are you sure you want to delete service %q?", c.Service)) {
-			return nil
-		}
+
+	ctx := gContext
+	client, err := switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, c.Zone)
+	if err != nil {
+		return err
 	}
 
-	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
-
-	nlb, err := globalstate.EgoscaleClient.FindNetworkLoadBalancer(ctx, c.Zone, c.NetworkLoadBalancer)
+	nlbs, err := client.ListLoadBalancers(ctx)
+	if err != nil {
+		return err
+	}
+	nlb, err := nlbs.FindLoadBalancer(c.NetworkLoadBalancer)
 	if err != nil {
 		return err
 	}
 
 	for _, service := range nlb.Services {
-		if *service.ID == c.Service || *service.Name == c.Service {
-			s := service
-			decorateAsyncOperation(fmt.Sprintf("Deleting service %q...", c.Service), func() {
-				err = globalstate.EgoscaleClient.DeleteNetworkLoadBalancerService(ctx, c.Zone, nlb, s)
-			})
-			if err != nil {
-				return err
+		if service.ID.String() == c.Service || service.Name == c.Service {
+
+			if !c.Force {
+				if !askQuestion(fmt.Sprintf("Are you sure you want to delete service %q?", service.ID)) {
+					return nil
+				}
 			}
 
-			return nil
+			op, err := client.DeleteLoadBalancerService(ctx, nlb.ID, service.ID)
+			decorateAsyncOperation(fmt.Sprintf("Deleting service %q...", c.Service), func() {
+				_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
+			})
+			return err
 		}
 	}