+++ cmd/nlb_service_update.go
@@ -4,16 +4,12 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/spf13/cobra"
 
-	"github.com/exoscale/cli/pkg/account"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
-	"github.com/exoscale/cli/utils"
-	egoscale "github.com/exoscale/egoscale/v2"
-	exoapi "github.com/exoscale/egoscale/v2/api"
+	v3 "github.com/exoscale/egoscale/v3"
 )
 
 type nlbServiceUpdateCmd struct {
@@ -24,20 +20,20 @@ type nlbServiceUpdateCmd struct {
 	NetworkLoadBalancer string `cli-arg:"#" cli-usage:"LOAD-BALANCER-NAME|ID"`
 	Service             string `cli-arg:"#" cli-usage:"SERVICE-NAME|ID"`
 
-	Description         string `cli-usage:"service description"`
-	HealthcheckInterval int64  `cli-usage:"service health checking interval in seconds"`
-	HealthcheckMode     string `cli-usage:"service health checking mode (tcp|http|https)"`
-	HealthcheckPort     int64  `cli-usage:"service health checking port"`
-	HealthcheckRetries  int64  `cli-usage:"service health checking retries"`
-	HealthcheckTLSSNI   string `cli-flag:"healthcheck-tls-sni" cli-usage:"service health checking server name to present with SNI in https mode"`
-	HealthcheckTimeout  int64  `cli-usage:"service health checking timeout in seconds"`
-	HealthcheckURI      string `cli-usage:"service health checking URI (required in http(s) mode)"`
-	Name                string `cli-usage:"service name"`
-	Port                int64  `cli-usage:"service port"`
-	Protocol            string `cli-usage:"service network protocol (tcp|udp)"`
-	Strategy            string `cli-usage:"load balancing strategy (round-robin|source-hash)"`
-	TargetPort          int64  `cli-usage:"port to forward traffic to on target instances"`
-	Zone                string `cli-short:"z" cli-usage:"Network Load Balancer zone"`
+	Description         string      `cli-usage:"service description"`
+	HealthcheckInterval int64       `cli-usage:"service health checking interval in seconds"`
+	HealthcheckMode     string      `cli-usage:"service health checking mode (tcp|http|https)"`
+	HealthcheckPort     int64       `cli-usage:"service health checking port"`
+	HealthcheckRetries  int64       `cli-usage:"service health checking retries"`
+	HealthcheckTLSSNI   string      `cli-flag:"healthcheck-tls-sni" cli-usage:"service health checking server name to present with SNI in https mode"`
+	HealthcheckTimeout  int64       `cli-usage:"service health checking timeout in seconds"`
+	HealthcheckURI      string      `cli-usage:"service health checking URI (required in http(s) mode)"`
+	Name                string      `cli-usage:"service name"`
+	Port                int64       `cli-usage:"service port"`
+	Protocol            string      `cli-usage:"service network protocol (tcp|udp)"`
+	Strategy            string      `cli-usage:"load balancing strategy (round-robin|source-hash)"`
+	TargetPort          int64       `cli-usage:"port to forward traffic to on target instances"`
+	Zone                v3.ZoneName `cli-short:"z" cli-usage:"Network Load Balancer zone"`
 }
 
 func (c *nlbServiceUpdateCmd) cmdAliases() []string { return nil }
@@ -57,128 +53,133 @@ func (c *nlbServiceUpdateCmd) cmdPreRun(cmd *cobra.Command, args []string) error
 }
 
 func (c *nlbServiceUpdateCmd) cmdRun(cmd *cobra.Command, _ []string) error {
-	var (
-		service *egoscale.NetworkLoadBalancerService
-		updated bool
-	)
 
-	ctx := exoapi.WithEndpoint(gContext, exoapi.NewReqEndpoint(account.CurrentAccount.Environment, c.Zone))
+	ctx := gContext
 
-	nlb, err := globalstate.EgoscaleClient.FindNetworkLoadBalancer(ctx, c.Zone, c.NetworkLoadBalancer)
+	client, err := switchClientZoneV3(ctx, globalstate.EgoscaleV3Client, c.Zone)
 	if err != nil {
 		return err
 	}
 
+	var updated bool
+
+	nlbs, err := client.ListLoadBalancers(ctx)
+	if err != nil {
+		return err
+	}
+	nlb, err := nlbs.FindLoadBalancer(c.NetworkLoadBalancer)
+	if err != nil {
+		return err
+	}
+
+	var service *v3.LoadBalancerService
 	for _, s := range nlb.Services {
-		if *s.ID == c.Service || *s.Name == c.Service {
-			service = s
-			break
+		if c.Service == string(s.ID) || c.Service == s.Name {
+			service = &s
 		}
 	}
 	if service == nil {
 		return errors.New("service not found")
 	}
 
+	svc := v3.UpdateLoadBalancerServiceRequest{
+		Healthcheck: service.Healthcheck,
+	}
+
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Description)) {
-		service.Description = &c.Description
+		svc.Description = c.Description
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckInterval)) {
-		hcInterval := time.Duration(c.HealthcheckInterval) * time.Second
-		service.Healthcheck.Interval = &hcInterval
+		svc.Healthcheck.Interval = c.HealthcheckInterval
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckMode)) {
-		service.Healthcheck.Mode = &c.HealthcheckMode
+		svc.Healthcheck.Mode = v3.LoadBalancerServiceHealthcheckMode(c.HealthcheckMode)
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckPort)) {
-		hcPort := uint16(c.HealthcheckPort)
-		service.Healthcheck.Port = &hcPort
+		svc.Healthcheck.Port = c.HealthcheckPort
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckRetries)) {
-		service.Healthcheck.Retries = &c.HealthcheckRetries
-		updated = true
-	}
-
-	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckTLSSNI)) {
-		service.Healthcheck.TLSSNI = &c.HealthcheckTLSSNI
+		svc.Healthcheck.Retries = c.HealthcheckRetries
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckTimeout)) {
-		hcTimeout := time.Duration(c.HealthcheckTimeout) * time.Second
-		service.Healthcheck.Timeout = &hcTimeout
+		svc.Healthcheck.Timeout = c.HealthcheckTimeout
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckURI)) {
-		service.Healthcheck.URI = &c.HealthcheckURI
+		svc.Healthcheck.URI = c.HealthcheckURI
+		updated = true
+	}
+	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.HealthcheckTLSSNI)) {
+		svc.Healthcheck.TlsSNI = c.HealthcheckTLSSNI
 		updated = true
 	}
 
 	// If mode is is tcp, ensure URI and TLSSNI are not set
-	if *service.Healthcheck.Mode == "tcp" && (utils.DefaultString(service.Healthcheck.TLSSNI, "") != "" || utils.DefaultString(service.Healthcheck.URI, "") != "") {
-		service.Healthcheck = &egoscale.NetworkLoadBalancerServiceHealthcheck{
-			Interval: service.Healthcheck.Interval,
-			Mode:     service.Healthcheck.Mode,
-			Port:     service.Healthcheck.Port,
-			Retries:  service.Healthcheck.Retries,
-			Timeout:  service.Healthcheck.Timeout,
-		}
-		updated = true
+	if c.HealthcheckMode == "tcp" && c.HealthcheckTLSSNI != "" {
+		return fmt.Errorf("cannot setup healthcheck TLSSNI with TCP mode (current value: %q)", c.HealthcheckTLSSNI)
+	}
+	if c.HealthcheckMode == "tcp" && c.HealthcheckURI != "" {
+		return fmt.Errorf("cannot setup healthcheck URI with TCP mode (current value: %q)", c.HealthcheckURI)
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Name)) {
-		service.Name = &c.Name
+		svc.Name = c.Name
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Port)) {
-		port := uint16(c.Port)
-		service.Port = &port
+		svc.Port = c.Port
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Protocol)) {
-		service.Protocol = &c.Protocol
+		svc.Protocol = v3.UpdateLoadBalancerServiceRequestProtocol(c.Protocol)
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.Strategy)) {
-		service.Strategy = &c.Strategy
+		svc.Strategy = v3.UpdateLoadBalancerServiceRequestStrategy(c.Strategy)
 		updated = true
 	}
 
 	if cmd.Flags().Changed(mustCLICommandFlagName(c, &c.TargetPort)) {
-		targetPort := uint16(c.TargetPort)
-		service.TargetPort = &targetPort
+		svc.TargetPort = c.TargetPort
 		updated = true
 	}
 
-	decorateAsyncOperation(fmt.Sprintf("Updating service %q...", c.Service), func() {
-		if updated {
-			if err = globalstate.EgoscaleClient.UpdateNetworkLoadBalancerService(ctx, c.Zone, nlb, service); err != nil {
-				return
-			}
+	if updated {
+		op, err := client.UpdateLoadBalancerService(ctx, nlb.ID, service.ID, svc)
+		if err != nil {
+			return err
 		}
-	})
-	if err != nil {
-		return err
-	}
 
-	if !globalstate.Quiet {
-		return (&nlbServiceShowCmd{
-			cliCommandSettings:  c.cliCommandSettings,
-			NetworkLoadBalancer: *nlb.ID,
-			Service:             *service.ID,
-			Zone:                c.Zone,
-		}).cmdRun(nil, nil)
+		decorateAsyncOperation(fmt.Sprintf("Updating service %q...", c.Service), func() {
+			_, err = client.Wait(ctx, op, v3.OperationStateSuccess)
+		})
+		if err != nil {
+			return err
+		}
+
+		if !globalstate.Quiet {
+			return (&nlbServiceShowCmd{
+				cliCommandSettings:  c.cliCommandSettings,
+				NetworkLoadBalancer: nlb.ID.String(),
+				Service:             service.ID.String(),
+				Zone:                c.Zone,
+			}).cmdRun(nil, nil)
+		}
+
 	}
 
 	return nil